Inline turn animation and remove dynamic import
